// nest-api-gateway-auth/generated/.proto/User.ts
/**
* This file is auto-generated by nestjs-proto-gen-ts
*/

import { Observable } from 'rxjs';
import { Metadata } from '@grpc/grpc-js';

export namespace com {
    export namespace ecommerce {
        export namespace aspnet {
            export namespace user {
                export interface UserService {
                    getUserById(
                        data: GetUserByIdRequest,
                        metadata?: Metadata,
                        ...rest: any[]
                    ): Observable<UserProfile>;
                    createUser(
                        data: CreateUserRequest,
                        metadata?: Metadata,
                        ...rest: any[]
                    ): Observable<UserProfile>;
                    getUserByEmailAndPassword(
                        data: GetUserByEmailAndPasswordRequest,
                        metadata?: Metadata,
                        ...rest: any[]
                    ): Observable<UserProfile>;
                    changePassword(
                        data: ChangePasswordRequest,
                        metadata?: Metadata,
                        ...rest: any[]
                    ): Observable<UserProfile>;
                }
                export interface ChangePasswordRequest {
                    id?: google.protobuf.StringValue;
                    oldPassword?: google.protobuf.StringValue;
                    newPassword?: google.protobuf.StringValue;
                }
                export interface GetUserByIdRequest {
                    id?: google.protobuf.StringValue;
                }
                export interface GetUserByEmailAndPasswordRequest {
                    email?: google.protobuf.StringValue;
                    password?: google.protobuf.StringValue;
                }
                export interface CreateUserRequest {
                    name?: google.protobuf.StringValue;
                    email?: google.protobuf.StringValue;
                    password?: google.protobuf.StringValue;
                }
                export interface UserProfile {
                    id?: string;
                    name?: string;
                    email?: string;
                }
            }
        }
    }
}
export namespace google {
    export namespace protobuf {
        export interface DoubleValue {
            value?: number;
        }
        export interface FloatValue {
            value?: number;
        }
        export interface Int64Value {
            value?: number;
        }
        export interface UInt64Value {
            value?: number;
        }
        export interface Int32Value {
            value?: number;
        }
        export interface UInt32Value {
            value?: number;
        }
        export interface BoolValue {
            value?: boolean;
        }
        export interface StringValue {
            value?: string;
        }
        export interface BytesValue {
            value?: Uint8Array;
        }
    }
}